+++ dict_plus/indexes/index.py
@@ -1,8 +1,85 @@
-from dict_plus.etypes import *
-import zlib
+class Index(object):
+    """
+    Index object to keep track of data internals
+    """
+
+    def make_hash(self, o):
+        raise NotImplementedError
+
+    def __init__(self, data=None):
+        """Create a new Iterable Index
+
+        Args:
+            data: Internal data dict to create the index from, optional
+
+        """
+        self.__data = {} if not data else data.copy()
+
+    def get(self, key):
+        """Get a value from the index
+
+        Args:
+            key: Key to get the value of
+
+        Returns:
+            Integer index of the key's location in the element list
+
+        """
+        raise NotImplementedError
+
+    def set(self, key, value):
+        """Set a key's location in the index
+
+        Args:
+            key: Key to get the location of
+            value: Integer value to set in the index
+
+        """
+        raise NotImplementedError
+
+    def has(self, key):
+        """Check whether the index has a given key in it
+
+        Args:
+            key: Key to check for
+
+        Returns:
+            True if the key exists, else False
+
+        """
+        raise NotImplementedError
+
+    def pop(self, key):
+        """ Remove and get the value of the given key
+
+        Args:
+            key: Key to get the value of
+
+        Returns:
+            Integer index value of the key in the element list
+
+        """
+        raise NotImplementedError
+
+    def isempty(self):
+        """Check whether the index is empty
+
+        Returns:
+            True if index is empty else False
+        """
+        raise NotImplementedError
+
+    def copy(self):
+        """Copy this index
+
+        Returns:
+            A copy of the index
+
+        """
+        raise NotImplementedError
 
 
-class IterableIndex(object):
+class IterableIndex(Index):
     """Index object to keep track of 'unhashable' types
     """
 
@@ -44,6 +121,7 @@ class IterableIndex(object):
             data: Internal data dict to create the index from, optional
 
         """
+        super(IterableIndex, self).__init__(data)
         self.__data = {} if not data else data.copy()
 
     def get(self, key):
@@ -120,36 +198,31 @@ class IterableIndex(object):
         return self.__class__(self.__data)
 
 
-class SortedIterableIndex(IterableIndex):
-    """Index object to keep track of 'unhashable' types
+class NullIndex(Index):
     """
+    Index that does nothing
+    """
+    def __init__(self, data=None):
+        super(NullIndex, self).__init__()
+        self.__data = None
+
     def make_hash(self, o):
-        """Makes a hash for a given object, doesn't guarantee collisions won't happen.
+        return 0
 
-        Args:
-            o: Object to get a hash for
+    def get(self, key):
+        return None
 
-        Returns:
-            The hash of the object
+    def set(self, key, value):
+        pass
 
-        """
+    def has(self, key):
+        return False
 
-        if isinstance(o, StringTypes):
-            if six.PY3:
-                o = o.encode()
+    def pop(self, key):
+        return None
 
-            return zlib.crc32(o) & 0xffffffff
+    def isempty(self):
+        return True
 
-        if isinstance(o, list):
-            hashes = []
-            for el in o:
-                hashes.append(self.make_hash(el))
-            return self.make_hash(str(hashes) + "l")
-        elif isinstance(o, dict):
-            return self.make_hash(o.items())
-        elif isinstance(o, set):
-            return self.make_hash(str(self.make_hash(list(o))) + "s")
-        elif hasattr(o, "__str__"):
-            return self.make_hash(str(o) + "_s")
-        else:
-            raise TypeError("Can't hash {}, submit an issue!".format(o))
+    def copy(self):
+        return self.__class__(self.__data)