+++ tests/test_listplus.py
@@ -1,0 +1,7 @@
+from dict_plus.lists import ListPlus
+
+
+def test_all():
+    l = ListPlus()
+    l.insert(0, 5)
+    tw = 2